+++ starkbank/brcodepreview/__init__.py
@@ -1,1 +1,0 @@
-from .__brcodepreview import query