+++ starkbank/taxpayment/__taxpayment.py
@@ -17,14 +17,14 @@ class TaxPayment(Resource):
     - scheduled [datetime.date or string, default today]: payment scheduled date. ex: datetime.date(2020, 3, 10)
     - tags [list of strings]: list of strings for tagging
     ## Attributes (return-only):
-    - id [string, default None]: unique id returned when payment is created. ex: "5656565656565656"
-    - type [string, default None]: tax type. ex: "das"
-    - status [string, default None]: current payment status. ex: "success" or "failed"
-    - amount [int, default None]: amount automatically calculated from line or bar_code. ex: 23456 (= R$ 234.56)
-    - fee [integer, default None]: fee charged when tax payment is created. ex: 200 (= R$ 2.00)
+    - id [string]: unique id returned when payment is created. ex: "5656565656565656"
+    - type [string]: tax type. ex: "das"
+    - status [string]: current payment status. ex: "success" or "failed"
+    - amount [int]: amount automatically calculated from line or bar_code. ex: 23456 (= R$ 234.56)
+    - fee [integer]: fee charged when tax payment is created. ex: 200 (= R$ 2.00)
     - transaction_ids [list of strings]: ledger transaction ids linked to this TaxPayment. ex: ["19827356981273"]
-    - updated [datetime.datetime, default None]: latest update datetime for the payment. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
-    - created [datetime.datetime, default None]: creation datetime for the payment. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
+    - updated [datetime.datetime]: latest update datetime for the payment. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
+    - created [datetime.datetime]: creation datetime for the payment. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
     """
 
     def __init__(self, description, line=None, bar_code=None, tags=None, scheduled=None, id=None, type=None, amount=None,