+++ starkbank/webhook/__webhook.py
@@ -10,8 +10,8 @@ class Webhook(Resource):
     ## Parameters (required):
     - url [string]: Url that will be notified when an event occurs.
     - subscriptions [list of strings]: list of any non-empty combination of the available services. ex: ["transfer", "invoice", "deposit"]
-    ## Attributes:
-    - id [string, default None]: unique id returned when the webhook is created. ex: "5656565656565656"
+    ## Attributes (return-only):
+    - id [string]: unique id returned when the webhook is created. ex: "5656565656565656"
     """
 
     def __init__(self, url, subscriptions, id=None):